+++ package.json
@@ -3,14 +3,55 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./server/bin/www",
+    "dev": "nodemon --exec babel-node ./server/bin/www",
+    "devb": "babel-watch ./server/bin/www",
+    "clean": "rm -rf dist",
+    "build": "npm run clean && babel ./server --out-dir dist --copy-files",
+    "itgam": "ls"
   },
   "dependencies": {
+    "app-root-path": "^3.0.0",
     "cookie-parser": "~1.4.4",
+    "cross-env": "^7.0.3",
     "debug": "~2.6.9",
+    "dotenv": "^10.0.0",
     "express": "~4.16.1",
-    "hbs": "~4.0.4",
+    "express-handlebars": "^5.3.2",
+    "hbs": "^4.1.1",
     "http-errors": "~1.6.3",
-    "morgan": "~1.9.1"
+    "mongoose": "^6.0.5",
+    "morgan": "~1.9.1",
+    "rimraf": "^3.0.2",
+    "webpack": "^5.35.0",
+    "webpack-dev-middleware": "^4.2.0",
+    "webpack-hot-middleware": "^2.25.0",
+    "winston": "^3.3.3",
+    "yup": "^0.32.11"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.13.14",
+    "@babel/core": "^7.13.15",
+    "@babel/node": "^7.13.13",
+    "@babel/preset-env": "^7.13.15",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-module-resolver": "^4.1.0",
+    "babel-preset-minify": "^0.5.1",
+    "babel-watch": "^7.4.1",
+    "core-js": "^3.11.1",
+    "css-loader": "^5.2.4",
+    "eslint": "^7.26.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-import-resolver-babel-module": "^5.3.1",
+    "eslint-plugin-import": "^2.23.2",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-webpack-plugin": "^3.1.1",
+    "mini-css-extract-plugin": "^1.6.0",
+    "nodemon": "^2.0.19",
+    "prettier": "^2.4.1",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^4.11.0"
   }
 }
