+++ public/vanesa.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+
+<body>
+    <h1>Mi Pagina Creeper</h1>
+    <img src="C:/Users/Vanesa R/AppData/Local/Packages/CanonicalGroupLimited.Ubuntu20.04LTS_79rhkp1fndgsc/LocalState/rootfs/home/vaneriano/pwpc2/projnotes2022b/public/images/Creeper.jpg"
+        width="100" alt="Creeper Image">
+</body>
+
+</html>
