+++ server/bin/www
@@ -3,8 +3,10 @@
 /**
  * Module dependencies.
  */
-
-var app = require('../app');
+// ES5 
+// var app = require('../app');
+// ES6 
+import app from "../app"
 var debug = require('debug')('projnotes2022b:server');
 var http = require('http');
 
@@ -19,6 +21,8 @@ app.set('port', port);
  * Create HTTP server.
  */
 
+// app es una funcion de tipo middleware (codigo intermediario)
+// (req, res) =>{ ...res.send("algo") }
 var server = http.createServer(app);
 
 /**
@@ -59,7 +63,8 @@ function onError(error) {
   }
 
   var bind = typeof port === 'string'
-    ? 'Pipe ' + port
+    // ? 'Pipe ' + port
+    ? `Pipe ${port}`
     : 'Port ' + port;
 
   // handle specific listen errors with friendly messages
@@ -87,4 +92,7 @@ function onListening() {
     ? 'pipe ' + addr
     : 'port ' + addr.port;
   debug('Listening on ' + bind);
+  // Desestrecuturando port de addr
+  let { port } = addr
+  console.log(`Listening at http://localhost:${port}`);
 }
